feat: fetch and display popular
