+++ src/api/getPopular.ts
@@ -2,7 +2,7 @@ import { baseUrl } from 'constants/baseUrl';
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
 import { TTag } from 'types/types';
 
-export const getPopular = createApi({
+const getPopular = createApi({
   reducerPath: 'getPopular',
   baseQuery: fetchBaseQuery({
     baseUrl,
@@ -11,7 +11,10 @@ export const getPopular = createApi({
     fetchPopularImages: build.query <TTag[], number | string>({
       query: () => ({
         url: 'tags/',
+        credentials: 'include', 
       }),
     }),
   }),
 });
+
+export { getPopular };
