+++ src/components/Forms/AuthorizationForm/AuthorizationForm.tsx
@@ -48,10 +48,11 @@ const AuthorizationForm = () => {
   return (
     <form onSubmit={authUserRequest} className={styles.container}>
       <h1 className={styles.title}>Авторизация</h1>
-
+      
+      {false && (
       <YandexLogin clientID={clientID}>
         <LoginWithButton title='Войти с Яндекс ID' icon={<YandexIcon />} />
-      </YandexLogin>
+      </YandexLogin>)}
 
       <LineWithWord text='Или' />
