--- src/constants/popularTestData.ts
@@ -1,122 +1,0 @@
-export const popularData = [
-  {
-      "id": 8,
-      "tag_images": [
-          {
-              "id": 70,
-              "name": "https://i.pinimg.com/736x/06/7b/82/067b828091cd4ead12c986311abc1d78.jpg",
-              "image": "/media/images/067b828091cd4ead12c986311abc1d78.jpg"
-          }
-      ],
-      "name": "Спорт",
-      "slug": "sport"
-  },
-  {
-      "id": 9,
-      "tag_images": [
-          {
-              "id": 146,
-              "name": "https://i.pinimg.com/550x/95/9a/19/959a19fe9c32cac0c7584ddf9b66c7ed.jpg",
-              "image": "/media/images/959a19fe9c32cac0c7584ddf9b66c7ed_pnLD6wB.jpg"
-          }
-      ],
-      "name": "Развлечения",
-      "slug": "entertainment"
-  },
-  {
-      "id": 2,
-      "tag_images": [
-          {
-              "id": 136,
-              "name": "https://i.pinimg.com/736x/be/93/b6/be93b6855af01d003287f11762146daf.jpg",
-              "image": "/media/images/be93b6855af01d003287f11762146daf.jpg"
-          }
-      ],
-      "name": "Путешествия",
-      "slug": "trevels"
-  },
-  {
-      "id": 1,
-      "tag_images": [
-          {
-              "id": 60,
-              "name": "https://i.pinimg.com/originals/f9/47/9d/f9479d4e3f0822838424bcb1854935f9.jpg",
-              "image": "/media/images/f9479d4e3f0822838424bcb1854935f9.jpg"
-          }
-      ],
-      "name": "Природа",
-      "slug": "nature"
-  },
-  {
-      "id": 10,
-      "tag_images": [
-          {
-              "id": 26,
-              "name": "https://i.pinimg.com/1200x/02/64/fa/0264faed0d6efafd3162be85cbdedc7d.jpg",
-              "image": "/media/images/0264faed0d6efafd3162be85cbdedc7d.jpg"
-          }
-      ],
-      "name": "Портрет",
-      "slug": "portret"
-  },
-  {
-      "id": 7,
-      "tag_images": [
-          {
-              "id": 141,
-              "name": "https://i.pinimg.com/564x/12/39/b5/1239b5b64faeab9b9306edac5603da2f.jpg",
-              "image": "/media/images/1239b5b64faeab9b9306edac5603da2f.jpg"
-          }
-      ],
-      "name": "Музыка",
-      "slug": "music"
-  },
-  {
-      "id": 5,
-      "tag_images": [
-          {
-              "id": 5,
-              "name": "https://i.pinimg.com/originals/49/50/90/495090c194d5ab41e0514c5e06f8facd.gif",
-              "image": "/media/images/495090c194d5ab41e0514c5e06f8facd.gif"
-          }
-      ],
-      "name": "Искусство",
-      "slug": "art"
-  },
-  {
-      "id": 3,
-      "tag_images": [
-          {
-              "id": 22,
-              "name": "https://i.pinimg.com/474x/78/9a/cf/789acf5bea5139bdbfc8f7b00f281b30.jpg",
-              "image": "/media/images/789acf5bea5139bdbfc8f7b00f281b30_ud4K2bK.jpg"
-          }
-      ],
-      "name": "Здоровье",
-      "slug": "health"
-  },
-  {
-      "id": 6,
-      "tag_images": [
-          {
-              "id": 43,
-              "name": "https://i.pinimg.com/736x/c3/6c/2f/c36c2fa7e265a9965197c6978cd1721a.jpg",
-              "image": "/media/images/c36c2fa7e265a9965197c6978cd1721a.jpg"
-          }
-      ],
-      "name": "Еда",
-      "slug": "food"
-  },
-  {
-      "id": 4,
-      "tag_images": [
-          {
-              "id": 84,
-              "name": "https://i.pinimg.com/736x/b9/76/a7/b976a74432f46bb4aee6b907933a15df.jpg",
-              "image": "/media/images/b976a74432f46bb4aee6b907933a15df.jpg"
-          }
-      ],
-      "name": "Город",
-      "slug": "city"
-  }
-];
